LS-60099 addressing code warnings.  (#69)
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Cleaning up some repo warnings to address:
commenting
removing unused parameters `host` and `topoinline`
changing a slice declaration to nil
renamed `min `and `max` to `minimum` and` maximum` to eliminate the
collision warnings with builtin functions

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

Behavior is the same, just cleaned up some warnings

---

## Pull request checklist

Please check if your PR fulfills the following requirements:

- [x] Tests(`make test`) for the changes have been added (for bug fixes
/ features) and pass
- [ ] Docs have been reviewed and added / updated if needed (for bug
fixes / features)
- [x] Lint (`make lint`) has passed locally and any fixes were made for
failures

## Pull request type

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [x] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

I was going to update the deprecated rand.Seed usage but after starting
this I think it should be it's own PR due to the amount of work it will
be. We have LS-60180 for TODO comments.

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->