[LS-60884] remove uses of rand.Seed (#71)
## What is the current behavior?
`rand.Seed` is deprecated, and this deprecated method being in use is
preventing the linter from allowing changes to this repo


## What is the new behavior?
All uses of `rand.Seed(someseed)` are removed. This is replaced with
`rand.New(rand.NewSource(someseed))` which is the preferred way to
handle this according to doc strings. Instead of using the global `rand`
in random generation, the preferred method seems to be to pass down the
generator created by `rand.New` so this also passed down generators into
any method that uses them

There is a question here of where we should pass generators as function
args and where we should just put the generator as a field in the
structs, but a bunch of these structs have json annotations that suggest
they are read from files and i do not really know how to make sure
`rand` gets populated in the case where some of these structs are
constructed from json

---------

Co-authored-by: jason.crouse <jason.crouse@servicenow.com>